kin-statistics/requirements.txt
absl-py==1.3.0
amqp==5.1.1
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
DAWG-Python==0.7.2
dependency-injector==4.40.0
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
docopt==0.6.2
drf-yasg==1.21.4
exceptiongroup==1.0.4
Faker==15.3.3
flake8==6.0.0
flatbuffers==22.10.26
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.11.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kombu==5.2.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
nltk==3.7
np==1.0.2
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
prompt-toolkit==3.0.32
protobuf==3.19.6
psycopg2-binary==2.9.4
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
PyJWT==2.5.0
pymongo==4.3.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.3
Telethon==1.25.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
